requirements.txt
anthropic>=0.34.0
requests>=2.31.0
pandas>=2.0.0
plotly>=5.18.0
dash>=2.14.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
openpyxl>=3.1.0
